@@ -4,21 +4,18 @@
   "description": "Tool to scrap Etsy API from web, then create typescript wrapper.",
   "main": "./js/main.js",
   "dependencies": {
-    "cheerio": "^0.22.0",
+    "cheerio": "^1.0.0-rc.3",
     "cheerio-tableparser": "^1.0.1",
     "lodash": "^4.17.4",
-    "node-fetch": "^1.7.3",
-    "request": "^2.81.0",
-    "request-promise": "^4.2.1"
+    "request": "^2.79.0",
+    "request-promise": "^4.1.1"
   },
   "devDependencies": {
-    "@types/cheerio": "^0.22.2",
+    "@types/cheerio": "^0.22.13",
     "@types/lodash": "^4.14.54",
-    "@types/prettier": "^1.7.0",
-    "@types/request": "^2.0.3",
-    "@types/request-promise": "^4.1.37",
-    "prettier": "^1.7.0",
-    "typescript": "^2.2.1"
+    "@types/request": "2.48.3",
+    "@types/request-promise": "^4.1.33",
+    "typescript": "^3.6.3"
   },
   "scripts": {
     "test": "mocha",
